backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
anthropic==0.28.0
pdfplumber==0.11.0
python-docx==1.1.2
reportlab==4.2.0
aiofiles==23.2.1
